Fix bug in TestDataChannelParamters
Check data channel label to make sure it is the one we want. There's a
weird requirement in the WASM bindings that you have to create a data
channel in order to establish the connection, so OnDataChannel was
getting triggered with that one instead of the one we intended to test.